+++ CHANGELOG.md
@@ -2,8 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+
+* Implemented the additional render options. You can now render ships with armour, hull, and system damage (including core systems).
+* Added a `facingArc` property to the turret system. When rendered, the turret's left-most arc is rotated to the `facingArc` position, and all contained weapon arcs are rotated the difference between `leftArc` and `facingArc`.
 
 ## [v2.1.0] - 2023-04-29